# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forma-te API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "moises-alberto@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/ebook-content": {
            "post": {
                "tags": [
                    "Ebook Contents"
                ],
                "summary": "Create a new EbookContent",
                "description": "Creates a new EbookContent using the provided data",
                "operationId": "createEbookContent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateEbookContentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "EbookContent created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EbookContentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ebook/{ebookId}/ebook-content": {
            "get": {
                "tags": [
                    "Ebook Contents"
                ],
                "summary": "Get modules by EbookContent ID",
                "description": "Returns the EbookContent associated with a given course ID",
                "operationId": "getContentByEbookId",
                "parameters": [
                    {
                        "name": "ebookId",
                        "in": "path",
                        "description": "The ID of the ebook",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modules retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EbookContentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ebook-contents/{id}": {
            "put": {
                "tags": [
                    "Ebook Contents"
                ],
                "summary": "Update an ebook content",
                "description": "Update an existing ebook content by ID",
                "operationId": "updateEbookContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the ebook content to update.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateEbookContentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ebook content updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EbookContentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate entry for the ebook content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Duplicate entry for the ebook content"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred while updating the ebook content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred while updating the ebook content"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ebook-content/{id}": {
            "delete": {
                "tags": [
                    "Ebook Contents"
                ],
                "summary": "Delete a EbookContent",
                "description": "Deletes an existing EbookContent by its ID",
                "operationId": "destroyEbookContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the EbookContent to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "EbookContent deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreUpdateEbookContentRequest": {
                "required": [
                    "ebook_id",
                    "name"
                ],
                "properties": {
                    "ebook_id": {
                        "description": "The ID of the ebook",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the EbookContent",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the EbookContent",
                        "type": "string"
                    },
                    "file": {
                        "description": "The file of the EbookContent",
                        "type": "string"
                    },
                    "published": {
                        "description": "The published status of the EbookContent",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "EbookContentResource": {
                "title": "Ebook Content Resource",
                "description": "Resource representation of Ebook Content",
                "properties": {
                    "id": {
                        "description": "The ID of the ebook content",
                        "type": "integer"
                    },
                    "ebook_id": {
                        "description": "The ID of the associated ebook",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the ebook content",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the ebook content",
                        "type": "string"
                    },
                    "file": {
                        "description": "The file associated with the ebook content",
                        "type": "string"
                    },
                    "free": {
                        "description": "Indicates if the ebook content is free",
                        "type": "boolean"
                    },
                    "published": {
                        "description": "Indicates if the ebook content is published",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    }
}